add debug info

+++ src/ss_data.c
@@ -967,7 +967,7 @@ int ssd_copy_list(struct obj_data *to, struct list_head *od_list)
                     {
                         for (int j = 0; j < 8; j++)
                         {
-                            printf("%lf ",*(buffer+i*8+j));
+                            printf("%lf ",(double) *(buffer+i*8+j));
                         }
                         printf("\n");       
                     }
